+++ http/src/h2/acceptor/tests.rs
@@ -16,6 +16,7 @@ mod client;
 mod fixture;
 mod flow_control;
 mod lifecycle;
+mod recv_headers;
 mod send;
 mod shutdown;
 mod two_driver;